refactor: NSVisualEffectViewBlur
