+++ SwiftDiffusion/Views/MainView.swift
@@ -100,25 +100,3 @@ extension NSTextView {
     
   }
 }
-
-
-import SwiftUI
-import AppKit
-
-struct VisualEffectBlurView: NSViewRepresentable {
-  var material: NSVisualEffectView.Material
-  var blendingMode: NSVisualEffectView.BlendingMode
-  
-  func makeNSView(context: Context) -> NSVisualEffectView {
-    let view = NSVisualEffectView()
-    view.material = material
-    view.blendingMode = blendingMode
-    view.state = .active
-    return view
-  }
-  
-  func updateNSView(_ nsView: NSVisualEffectView, context: Context) {
-    nsView.material = material
-    nsView.blendingMode = blendingMode
-  }
-}
