+++ SwiftDiffusion/Views/VisualEffectBlurView.swift
@@ -1,0 +1,27 @@
+//
+//  VisualEffectBlurView.swift
+//  SwiftDiffusion
+//
+//  Created by Justin Bush on 2/5/24.
+//
+
+import SwiftUI
+import AppKit
+
+struct VisualEffectBlurView: NSViewRepresentable {
+  var material: NSVisualEffectView.Material
+  var blendingMode: NSVisualEffectView.BlendingMode
+  
+  func makeNSView(context: Context) -> NSVisualEffectView {
+    let view = NSVisualEffectView()
+    view.material = material
+    view.blendingMode = blendingMode
+    view.state = .active
+    return view
+  }
+  
+  func updateNSView(_ nsView: NSVisualEffectView, context: Context) {
+    nsView.material = material
+    nsView.blendingMode = blendingMode
+  }
+}
